+++ src/main/java/com/paneedah/mwc/asm/Interceptors.java
@@ -1,5 +1,6 @@
 package com.paneedah.mwc.asm;
 
+import com.paneedah.mwc.gui.HUD;
 import com.paneedah.mwc.instancing.PlayerWeaponInstance;
 import com.paneedah.weaponlib.*;
 import com.paneedah.weaponlib.animation.AnimationModeProcessor;
@@ -743,7 +744,7 @@ public class Interceptors {
 
         if (ClientModContext.getContext() != null && ClientModContext.getContext().getMainHeldWeapon() != null) {
             PlayerWeaponInstance pwi = ClientModContext.getContext().getMainHeldWeapon();
-            if (CustomGui.isInModifyingState(pwi) || CustomGui.isInAltModifyingState(pwi)) {
+            if (HUD.isInModifyingState(pwi) || HUD.isInAltModifyingState(pwi)) {
                 yawDelta *= 0.01f;
                 pitchDelta *= 0.01f;
             }