+++ src/main/java/com/paneedah/mwc/equipment/Electronics.java
@@ -24,7 +24,6 @@ public class Electronics {
                     GL11.glTranslatef(-0.12f, 0.56f, 0.01f);
                 })
                 .withCreativeTab(MWC.EQUIPMENT_TAB)
-                .withCrosshair("HP")
                 .withModel(new TabletModel(), "IPad.png")
                 //.withModel(new com.paneedah.mwc.models.LPscope(), "HP2.png")
                 .withFirstPersonPositioning(() -> {
@@ -147,7 +146,6 @@ public class Electronics {
                     GL11.glTranslatef(0.017f, 0.16f, 0.17f);
                 })
                 .withCreativeTab(MWC.EQUIPMENT_TAB)
-                .withCrosshair("HP")
                 .withModel(new GasDetector(), "gasdetector.png")
                 .withFirstPersonPositioning(() -> {
                     GL11.glScalef(4f, 4f, 4f);