+++ src/main/java/com/paneedah/weaponlib/ItemAttachment.java
@@ -20,12 +20,9 @@ import java.util.Collections;
 import java.util.List;
 import java.util.function.Function;
 
-import static com.paneedah.mwc.ProjectConstants.ID;
-
 public class ItemAttachment<T> extends Item implements ModelSource, IModernCraftingRecipe {
 
     private final AttachmentCategory category;
-    private final String crosshair;
     private final ApplyHandler<T> apply;
     private final ApplyHandler<T> remove;
     protected ApplyHandler2<T> apply2;
@@ -65,22 +62,20 @@ public class ItemAttachment<T> extends Item implements ModelSource, IModernCraft
         void apply(ItemAttachment<T> itemAttachment, PlayerMeleeInstance instance);
     }
 
-    protected ItemAttachment(AttachmentCategory category, ModelBase model, String textureName, String crosshair,
+    protected ItemAttachment(AttachmentCategory category, ModelBase model, String textureName,
                              ApplyHandler<T> apply, ApplyHandler<T> remove) {
         this.category = category;
 //		if(model != null) {
 //			this.texturedModels.add(new Tuple<ModelBase, String>(model, textureName));
 //		}
         this.textureName = textureName.toLowerCase();
-        this.crosshair = crosshair != null ? ID + ":" + "textures/crosshairs/" + crosshair + ".png" : null;
         this.apply = apply;
         this.remove = remove;
     }
 
-    protected ItemAttachment(AttachmentCategory category, String crosshair,
+    protected ItemAttachment(AttachmentCategory category,
                              ApplyHandler<T> apply, ApplyHandler<T> remove) {
         this.category = category;
-        this.crosshair = crosshair != null ? ID + ":" + "textures/crosshairs/" + crosshair + ".png" : null;
         this.apply = apply;
         this.remove = remove;
     }
@@ -141,12 +136,12 @@ public class ItemAttachment<T> extends Item implements ModelSource, IModernCraft
         return this;
     }
 
-    public ItemAttachment(AttachmentCategory category, String crosshair) {
-        this(category, crosshair, (a, w, p) -> {}, (a, w, p) -> {});
+    public ItemAttachment(AttachmentCategory category) {
+        this(category, (a, w, p) -> {}, (a, w, p) -> {});
     }
 
-    public ItemAttachment(AttachmentCategory category, ModelBase attachment, String textureName, String crosshair) {
-        this(category, attachment, textureName, crosshair, (a, w, p) -> {}, (a, w, p) -> {});
+    public ItemAttachment(AttachmentCategory category, ModelBase attachment, String textureName) {
+        this(category, attachment, textureName, (a, w, p) -> {}, (a, w, p) -> {});
     }
 
     public AttachmentCategory getCategory() {
@@ -164,10 +159,6 @@ public class ItemAttachment<T> extends Item implements ModelSource, IModernCraft
      */
     public void setFirstModel(ItemAttachment<Weapon> model) {
         texturedModels.set(0, model.getTexturedModels().get(0));
-    }
-
-    public String getCrosshair() {
-        return crosshair;
     }
 
     public ApplyHandler<T> getApply() {