+++ src/main/java/com/paneedah/weaponlib/ItemScope.java
@@ -137,7 +137,7 @@ public class ItemScope extends ItemAttachment<Weapon> {
     private final Builder builder;
 
     private ItemScope(Builder builder) {
-        super(AttachmentCategory.SCOPE, builder.getModel(), builder.getTextureName(), null,
+        super(AttachmentCategory.SCOPE, builder.getModel(), builder.getTextureName(),
                 null, null);
         this.builder = builder;