+++ src/main/java/com/paneedah/weaponlib/electronics/ItemHandheld.java
@@ -51,7 +51,7 @@ public class ItemHandheld<T> extends ItemAttachment<T>
     private final Builder<T> builder;
 
     private ItemHandheld(Builder<T> builder) {
-        super(AttachmentCategory.SCOPE, builder.getModel(), builder.getTextureName(), null,
+        super(AttachmentCategory.SCOPE, builder.getModel(), builder.getTextureName(),
                 null, null);
         this.builder = builder;